Update README.md
+1 for the graphic! Some suggestions: 

Info to add to the intro/About section:
Ln 13: Why would someone use this--what will it solve for the user? Who is most likely to find this interesting — ie what types of developers or technologists, under what circumstances? What are a few use cases (where/when would someone use this)? What does replacing .find ... achieve (make it simple to understand, for a newbie)?
- Where did inspiration for this project come from/under what conditions? Is there anything else like it available? If yes, what does this do better/differently compared to those options? 
- How has using this project helped you/others?
- LN 18/list of features: which of these are most valuable? Can you segregate them into most-to-least powerful categories? 

Install/Manual Install directions: are these complete? To find out, sit with someone and see if they can get this running based on the info provided. :)

I'd make separate pages for the "All options" and "environment variables" sections, highlighting in this README the most 1-2 (for each section) need-to-know(s).

Are you accepting contributions? A Contributing.md with details on how to contribute could help.

Add a license section at the bottom and the text of the MIT. 

Thanks for submitting to FMRM!